requirements.txt
pytest-django==4.4.0
requests~=2.26.0
Pillow~=9.3.0
sorl-thumbnail==12.7.0
Django==3.0.7
djangorestframework==3.12.4
django-extensions==2.2.9
django-cors-headers==3.3.0